+++ pages/graph/index.jsx
@@ -2,13 +2,7 @@ import BarChart from "@/components/BarChart";
 import DoughnutGraph from "@/components/DoughnutGraph";
 
 function GraphPage() {
-  return (
-    <>
-      <div>
-        <BarChart />
-      </div>
-    </>
-  );
+  return <BarChart />;
 }
 
 export default GraphPage;
